separated each part of the accounts page into separate pages
