+++ client/src/helpers/routes.ts
@@ -12,4 +12,7 @@ export const ACCOUNT = "/account"
 export const REGISTER_USER = `${REGISTER}${USER}`
 export const REGISTER_ORG = `${REGISTER}/${ORGANIZATION}`
 export const ACCOUNT_CREATE_ORG = `${ACCOUNT}${ORGANIZATION}`
+export const ACCOUNT_CHANGE_PASSWORD = `${ACCOUNT}/change-password`
+export const ACCOUNT_JOIN_ORGANIZATION = `${ACCOUNT}/join-organization`
+export const ACCOUNT_NOTIFICATION_SETTINGS = `${ACCOUNT}/notification-settings`
 export const NOTIFICATIONS = "/notifications"
