+++ client/src/pages/account/AccountOrganization.tsx
@@ -19,15 +19,11 @@ export const AccountOrganization = () => {
 
 	return (
 		<div className = "tw-flex tw-flex-row">
-			<div className = "tw-w-1/4">	
-			</div>
 			<div className = "tw-flex tw-flex-col tw-gap-y-2">
 				<h1>Create Organization</h1>
 				<p>By creating an organization under this account, you will automatically be added as the admin of the new organization. </p>
 				{error && "status" in error ? (error.data.errors?.map((errorMessage: string, i: number) => <p className = "--text-alert" key = {`org_error_${i}`}>{errorMessage}</p>)) : null}
 				<OrganizationForm onSubmit={onSubmit}/>
-			</div>
-			<div className = "tw-w-1/4">	
 			</div>
 		</div>
 	)
