+++ client/src/pages/account/NotificationSettings.tsx
@@ -1,0 +1,8 @@
+import React from "react"
+import { UserNotificationTypeForm } from "../../components/forms/UserNotificationTypeForm"
+
+export const NotificationSettings = () => {
+	return (
+		<UserNotificationTypeForm/>
+	)
+}
